chore: rename openai
